renamed notification adapter module; moved payloads to be as part of the module
