+++ src/domain/communication/updates/updates.module.ts
@@ -2,7 +2,7 @@ import { AuthorizationModule } from '@core/authorization/authorization.module';
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { CommunicationAdapterModule } from '@services/platform/communication-adapter/communication-adapter.module';
-import { NotificationAdapterModule } from '@services/platform/notifications-adapter/notification.adapter.module';
+import { NotificationAdapterModule } from '@services/platform/notification-adapter/notification.adapter.module';
 import { RoomModule } from '../room/room.module';
 import { Updates } from './updates.entity';
 import { UpdatesResolverFields } from './updates.resolver.fields';
