+++ src/domain/communication/updates/updates.resolver.mutations.ts
@@ -18,8 +18,8 @@ import { SUBSCRIPTION_UPDATE_MESSAGE } from '@common/constants/providers';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
 import { LogContext } from '@common/enums';
 import { getRandomId } from '@src/common';
-import { NotificationInputUpdateSent } from '@services/platform/notifications-adapter/dto/notification.dto.input.update.sent';
-import { NotificationAdapter } from '@services/platform/notifications-adapter/notification.adapter';
+import { NotificationInputUpdateSent } from '@services/platform/notification-adapter/dto/notification.dto.input.update.sent';
+import { NotificationAdapter } from '@services/platform/notification-adapter/notification.adapter';
 
 @Resolver()
 export class UpdatesResolverMutations {
