+++ src/domain/community/user/user.resolver.mutations.ts
@@ -24,8 +24,8 @@ import { PreferenceService } from '@domain/common/preference';
 import { UpdateUserPreferenceInput } from './dto/user.dto.update.preference';
 import { PreferenceSetService } from '@domain/common/preference-set/preference.set.service';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
-import { NotificationInputUserRegistered } from '@services/platform/notifications-adapter/dto/notification.dto.input.user.registered';
-import { NotificationAdapter } from '@services/platform/notifications-adapter/notification.adapter';
+import { NotificationInputUserRegistered } from '@services/platform/notification-adapter/dto/notification.dto.input.user.registered';
+import { NotificationAdapter } from '@services/platform/notification-adapter/notification.adapter';
 
 @Resolver(() => IUser)
 export class UserResolverMutations {
