+++ src/services/domain/registration/registration.module.ts
@@ -5,7 +5,7 @@ import { OrganizationModule } from '@domain/community/organization/organization.
 import { RegistrationResolverMutations } from './registration.resolver.mutations';
 import { PreferenceSetModule } from '@domain/common/preference-set/preference.set.module';
 import { AuthorizationModule } from '@core/authorization/authorization.module';
-import { NotificationAdapterModule } from '@services/platform/notifications-adapter/notification.adapter.module';
+import { NotificationAdapterModule } from '@services/platform/notification-adapter/notification.adapter.module';
 
 @Module({
   imports: [
