+++ src/services/platform/notification-adapter/dto/notification.dto.input.base.ts
