+++ src/services/platform/notification-adapter/dto/notification.dto.input.callout.published.ts
