+++ src/services/platform/notification-adapter/dto/notification.dto.input.collaboration.interest.ts
