+++ src/services/platform/notification-adapter/dto/notification.dto.input.community.new.member.ts
