+++ src/services/platform/notification-adapter/dto/notification.dto.input.user.registered.ts
