+++ src/services/platform/notification-adapter/event-payloads/community.collaboration.interest.event.payload.ts
