+++ src/services/platform/notification-adapter/event-payloads/hub.payload.ts
